requirements.txt
asgiref==3.2.10
astroid==2.4.2
cffi==1.14.1
colorama==0.4.3
cryptography==3.0
Django==3.1
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==7.2.0
pycparser==2.20
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
